+++ packages/util/src/bn/min.spec.ts
@@ -6,7 +6,7 @@ import { BN, bnMin } from '.';
 describe('bnMin', (): void => {
   it('finds BN minimum', (): void => {
     expect(
-      bnMin(new BN(1), new BN(2), new BN(3))
+      bnMin(new BN(2), new BN(1), new BN(3))
     ).toEqual(new BN(1));
   });
 });