+++ packages/util/src/bn/util.ts
@@ -7,5 +7,11 @@ import { BN } from './bn';
 export function checkMaxMin (type: 'max' | 'min', items: BN[]): BN {
   assert(items.length >= 1, 'Must provide one or more BN arguments');
 
-  return items.reduce<BN>((a, v) => BN[type](a, v), items[0]);
+  let result = items[0];
+
+  for (const v of items) {
+    result = BN[type](result, v);
+  }
+
+  return result;
 }