+++ packages/networks/src/interfaces.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2021 @polkadot/networks authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Network, SubstrateNetwork } from './types';
+import type { KnownSubstrate, Network, SubstrateNetwork } from './types';
 
 import { knownGenesis, knownIcon, knownLedger, knownTestnet } from './defaults';
 import { knownSubstrate } from './substrate';
@@ -10,37 +10,54 @@ import { knownSubstrate } from './substrate';
 const UNSORTED = [0, 2, 42];
 const TESTNETS = ['testnet'];
 
-export const allNetworks = knownSubstrate.map((o): SubstrateNetwork => {
+function toExpanded (o: KnownSubstrate): SubstrateNetwork {
   const network = o.network || '';
   const nameParts = network.replace(/_/g, '-').split('-');
   const n = o as SubstrateNetwork;
 
+  // ledger additions
   n.slip44 = knownLedger[network];
   n.hasLedgerSupport = !!n.slip44;
 
+  // general items
   n.genesisHash = knownGenesis[network] || [];
   n.icon = knownIcon[network] || 'substrate';
+
+  // filtering
   n.isTestnet = !!knownTestnet[network] || TESTNETS.includes(nameParts[nameParts.length - 1]);
   n.isIgnored = n.isTestnet || (!(o.standardAccount && o.decimals && o.symbols) && o.prefix !== 42);
 
   return n;
-});
+}
+
+function filterSelectable ({ genesisHash, prefix }: Network): boolean {
+  return !!genesisHash.length || prefix === 42;
+}
+
+function filterNetworks (n: SubstrateNetwork): n is Network {
+  return !n.isIgnored && !!n.network;
+}
+
+function sortNetworks (a: Network, b: Network): number {
+  return UNSORTED.includes(a.prefix) === UNSORTED.includes(b.prefix)
+    ? 0
+    : UNSORTED.includes(a.prefix)
+      ? -1
+      : UNSORTED.includes(b.prefix)
+        ? 1
+        : a.displayName.localeCompare(b.displayName);
+}
+
+// This is all the Substrate networks with our additional information
+export const allNetworks = knownSubstrate.map(toExpanded);
 
 // The list of available/claimed prefixes
 //   - no testnets
 //   - we only include those where we have a standardAccount
 //   - sort by name, however we keep 0, 2, 42 first in the list
 export const availableNetworks = allNetworks
-  .filter((n): n is Network => !n.isIgnored && !!n.network)
-  .sort((a, b) =>
-    UNSORTED.includes(a.prefix) === UNSORTED.includes(b.prefix)
-      ? 0
-      : UNSORTED.includes(a.prefix)
-        ? -1
-        : UNSORTED.includes(b.prefix)
-          ? 1
-          : a.displayName.localeCompare(b.displayName)
-  );
+  .filter(filterNetworks)
+  .sort(sortNetworks);
 
 // A filtered list of those chains we have details about (genesisHashes)
-export const selectableNetworks = availableNetworks.filter((n) => n.genesisHash.length || n.prefix === 42);
+export const selectableNetworks = availableNetworks.filter(filterSelectable);