+++ packages/util/src/versionDetect.ts
@@ -56,24 +56,24 @@ function getVersionLength (all: { version: string }[]): number {
 /** @internal */
 function flattenInfos (all: PackageJson[]): string {
   const verLength = getVersionLength(all);
+  const stringify = ({ name, version }: PackageJson) =>
+    `\t${version.padEnd(verLength)}\t${name}`;
 
-  return all
-    .map(({ name, version }) => `\t${version.padEnd(verLength)}\t${name}`)
-    .join('\n');
+  return all.map(stringify).join('\n');
 }
 
 /** @internal */
 function flattenVersions (entry: VersionPath[]): string {
-  const all = entry.map((version: VersionPath | string): VersionPath =>
+  const toPath = (version: VersionPath | string): VersionPath =>
     isString(version)
       ? { version }
-      : version
-  );
+      : version;
+  const all = entry.map(toPath);
   const verLength = getVersionLength(all);
+  const stringify = ({ path, version }: VersionPath) =>
+    `\t${version.padEnd(verLength)}\t${(!path || path.length < 5) ? '<unknown>' : path}`;
 
-  return all
-    .map(({ path, version }) => `\t${version.padEnd(verLength)}\t${(!path || path.length < 5) ? '<unknown>' : path}`)
-    .join('\n');
+  return all.map(stringify).join('\n');
 }
 
 /** @internal */