+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changes:
 
 - Add `arrayZip` to combine a key + value array into entries-like
 - Assert `chainCode` inputs to `derive*`, protecting against misuse
+- Adjust base library usage of `for` vs `reduce/map`
 - Adjust testnet detection in `@polkadot/networks`
 - Bump Ledger libraries to latest versions